Change several schemas' price-related columns to double type

+++ data/study.sql
@@ -7,7 +7,7 @@ CREATE TABLE `study`(
   `market_id` int NOT NULL,
   `market_code` varchar(8) NOT NULL,
   `date` date NOT NULL,
-  `value` decimal(40,30) NOT NULL,
-  `value_2` decimal(40,30) NULL,
+  `value` double NOT NULL,
+  `value_2` double NULL,
   PRIMARY KEY (`id`)
 ) ENGINE=InnoDB AUTO_INCREMENT=1 DEFAULT CHARSET=utf8;
